+++ packages/backend/src/services/services.ts
@@ -81,6 +81,7 @@ export const projectService = new ProjectService({
     jobModel,
     emailClient,
     spaceModel,
+    sshKeyPairModel,
 });
 
 export const shareService = new ShareService({