+++ packages/frontend/src/components/ProjectAccess/index.tsx
@@ -1,7 +1,6 @@
 import { subject } from '@casl/ability';
-import { Anchor } from '@mantine/core';
+import { Anchor, Group, Text } from '@mantine/core';
 import { FC, useState } from 'react';
-import { Subtitle, SubtitleWrapper } from '../../pages/ProjectSettings.styles';
 import { useApp } from '../../providers/AppProvider';
 import { Can } from '../common/Authorization';
 import ProjectAccess from './ProjectAccess';
@@ -28,8 +27,8 @@ const ProjectUserAccess: FC<ProjectUserAccessProps> = ({ projectUuid }) => {
                 />
             ) : (
                 <>
-                    <SubtitleWrapper>
-                        <Subtitle>
+                    <Group position="apart">
+                        <Text color="dimmed">
                             Learn more about permissions in our{' '}
                             <Anchor
                                 role="button"
@@ -39,7 +38,8 @@ const ProjectUserAccess: FC<ProjectUserAccessProps> = ({ projectUuid }) => {
                             >
                                 docs
                             </Anchor>
-                        </Subtitle>
+                        </Text>
+
                         {!showProjectAccessCreate && (
                             <Can
                                 I={'manage'}
@@ -58,7 +58,7 @@ const ProjectUserAccess: FC<ProjectUserAccessProps> = ({ projectUuid }) => {
                                 />
                             </Can>
                         )}
-                    </SubtitleWrapper>
+                    </Group>
 
                     <ProjectAccess projectUuid={projectUuid} />
                 </>